+++ include/model/Generator.hpp
@@ -51,12 +51,12 @@ template<template<typename Target> class Decorator>
 class Generator : public std::enable_shared_from_this<Generator<Decorator>> {
  public:
   // Purely virtual methods
-  virtual Decorator<Symbol> drawSymbol(unsigned int pos,
-                                       unsigned int phase = 0,
+  virtual Decorator<Symbol> drawSymbol(size_t pos,
+                                       size_t phase = 0,
                                        const Sequence& context = {}) const = 0;
 
-  virtual Decorator<Sequence> drawSequence(unsigned int size,
-                                           unsigned int phase = 0) const = 0;
+  virtual Decorator<Sequence> drawSequence(size_t size,
+                                           size_t phase = 0) const = 0;
 
   virtual RandomNumberGeneratorPtr randomNumberGenerator() const = 0;
