+++ include/model/SimilarityBasedSequenceWeighting.hpp
@@ -60,10 +60,10 @@ class SimilarityBasedSequenceWeighting
 
   /*=============================[ CONSTRUCTORS ]=============================*/
 
-  SimilarityBasedSequenceWeighting(std::map<Sequence, unsigned int> counter,
+  SimilarityBasedSequenceWeighting(std::map<Sequence, size_t> counter,
                                    double normalizer,
-                                   unsigned int skip_offset,
-                                   unsigned int skip_length,
+                                   size_t skip_offset,
+                                   size_t skip_length,
                                    Sequence skip_sequence);
 
   /*============================[ STATIC METHODS ]============================*/
@@ -71,9 +71,9 @@ class SimilarityBasedSequenceWeighting
   // Trainer
   static SelfPtr train(TrainerPtr<Standard, Self> trainer,
                        standard_training_algorithm,
-                       unsigned int alphabet_size,
-                       unsigned int skip_offset,
-                       unsigned int skip_length,
+                       size_t alphabet_size,
+                       size_t skip_offset,
+                       size_t skip_length,
                        Sequence skip_sequence);
 
   /*==========================[ OVERRIDEN METHODS ]===========================*/
@@ -81,52 +81,52 @@ class SimilarityBasedSequenceWeighting
 
   // StandardEvaluator
   Probability evaluateSymbol(SEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(SEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // CachedEvaluator
   void initializeCache(CEPtr<Standard> evaluator,
-                       unsigned int phase) override;
+                       size_t phase) override;
   Probability evaluateSymbol(CEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(CEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // StandardGenerator
   Standard<Symbol> drawSymbol(SGPtr<Standard> generator,
-                              unsigned int pos,
-                              unsigned int phase,
+                              size_t pos,
+                              size_t phase,
                               const Sequence& context) const override;
   Standard<Sequence> drawSequence(SGPtr<Standard> generator,
-                                  unsigned int size,
-                                  unsigned int phase) const override;
+                                  size_t size,
+                                  size_t phase) const override;
 
   // SimpleSerializer
   void serialize(SSPtr serializer) override;
 
  private:
   // Instance variables
-  std::map<Sequence, unsigned int> _counter;
-  unsigned int _skip_offset;
-  unsigned int _skip_length;
+  std::map<Sequence, size_t> _counter;
+  size_t _skip_offset;
+  size_t _skip_length;
   Sequence _skip_sequence;
   double _normalizer;
 
   /*============================[ STATIC METHODS ]============================*/
 
   static double
-  calculate_normalizer(unsigned int skip_length,
-                       unsigned int skip_offset,
-                       unsigned int max_length,
-                       const std::map<Sequence, unsigned int>& counter,
-                       unsigned int alphabet_size);
+  calculate_normalizer(size_t skip_length,
+                       size_t skip_offset,
+                       size_t max_length,
+                       const std::map<Sequence, size_t>& counter,
+                       size_t alphabet_size);
 };
 
 }  // namespace model
